Decouple coinType from Bitcoin::Transaction
To support ForkId, the signature hash type used by Bitcoin forks
(BCH & BTG), a "forkId" argument is added to "getSignatureHash"
function in this commit. This was not needed before because BCH
uses a fork id of "0" while BTG is 79.

"Zcash::Transaction::getSignatureHash" is also affected because
it shares the same TransactionSigner calss with Bitcoin and
Bitcoin forks. The argument is marked as "unused" to avoid confusion.

Tested:
Unit tests